chore: split share social saga
